fix: codacy error
