+++ api/src/test/java/com/bbolab/gaonna/api/WithAccountSecurityContextFactory.java
@@ -1,8 +1,9 @@
 package com.bbolab.gaonna.api;
 
 import com.bbolab.gaonna.api.security.model.UserPrincipal;
-import com.bbolab.gaonna.api.util.MemberFactory;
+import com.bbolab.gaonna.core.domain.member.AuthProvider;
 import com.bbolab.gaonna.core.domain.member.Member;
+import com.bbolab.gaonna.core.domain.member.Role;
 import com.bbolab.gaonna.core.repository.MemberRepository;
 import lombok.RequiredArgsConstructor;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -22,7 +23,7 @@ public class WithAccountSecurityContextFactory implements WithSecurityContextFac
     public SecurityContext createSecurityContext(WithAccount account) {
         String name = account.value();
 
-        Member member = MemberFactory.createMemberStub(name);
+        Member member = createMemberStub(name);
         memberRepository.save(member);
 
         UserPrincipal userDetails = UserPrincipal.create(member);
@@ -32,5 +33,16 @@ public class WithAccountSecurityContextFactory implements WithSecurityContextFac
         context.setAuthentication(authentication);
         return context;
     }
+
+    public static Member createMemberStub(String name) {
+        return Member.builder()
+                .name(name)
+                .nickname(name)
+                .email(name + "@bbobbi.com")
+                .provider(AuthProvider.kakao)
+                .role(Role.ROLE_USER)
+                .build();
+    }
+
 
 }
